feat: use Round Emoji Badge and break away codewars blocks
